requirements.txt
flask
mysql-connector-python
pandas
openpyxl
SQLAlchemy
pymysql
cryptography
python-dotenv
pyodbc
firebase_admin
kafka-python